backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-firestore==2.13.2
firebase-admin==6.2.0
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pyjwt==2.8.0